MAINT: changes to orso.py
