+++ orsopy/fileio/base.py
@@ -527,3 +527,23 @@ def _nested_update(d, u):
         else:
             d = {k: u[k]}
     return d
+
+
+def _dict_diff(old, new):
+    # recursive find differences between two dictionaries
+    out = {}
+    for key, value in new.items():
+        if key in old:
+            if type(value) is dict:
+                diff = _dict_diff(old[key], value)
+                if diff == {}:
+                    continue
+                else:
+                    out[key] = diff
+            elif old[key] == value:
+                continue
+            else:
+                out[key] = value
+        else:
+            out[key] = value
+    return out
